修改     1. 细节调整

+++ apilib/src/main/java/com/guider/health/apilib/model/TokenInfo.java
@@ -14,7 +14,7 @@ public class TokenInfo implements Parcelable {
 
     private String token;
     private String refreshToken;
-    private long accountId;
+    private int accountId;
     private int expired;
 
     public String getToken() {
@@ -33,11 +33,11 @@ public class TokenInfo implements Parcelable {
         this.refreshToken = refreshToken;
     }
 
-    public long getAccountId() {
+    public int getAccountId() {
         return accountId;
     }
 
-    public void setAccountId(long accountId) {
+    public void setAccountId(int accountId) {
         this.accountId = accountId;
     }
 
@@ -58,7 +58,7 @@ public class TokenInfo implements Parcelable {
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeString(this.token);
         dest.writeString(this.refreshToken);
-        dest.writeLong(this.accountId);
+        dest.writeInt(this.accountId);
         dest.writeInt(this.expired);
     }
